requirements.txt
channels
daphne
geoip2
psycopg2-binary
channels-redis==3.3.1
Django>=2.1.3
black
djangorestframework>=3.9.0
djangorestframework-gis
django-environ
django-leaflet
django-filter
django-cors-headers
python-dotenv
Pillow
requests
Twisted[tls,http2]
whitenoise
djangorestframework_simplejwt
django-allauth==0.58.2
django-rest-auth==0.9.5
drf_yasg

